Move test_rolling/ewm.py to new cudf classic directory structure
